requirements.txt
langchain_openai
python-dotenv
langchain_community
streamlit
langchain_core
sentence_transformers
PyPDF2
langchain
langchain-anthropic
langchain-google-genai
langchain-google-firestore
firestore
chromadb
tiktoken
langchainhub
firecrawl-py
wikipedia
tavily-python